% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_code}
\alias{add_code}
\title{Add code
Append a new unit record to the existing data frame}
\usage{
add_code(codes_df, new_code, new_code_desc, codes_df_path)
}
\arguments{
\item{codes_df}{Existing codes data frame}

\item{new_code}{text name of a new code (single name only)}

\item{new_code_desc}{text description of the code}

\item{codes_df_path}{full path to the codes data frame}
}
\description{
Add code
Append a new unit record to the existing data frame
}
