% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{add_codes_to_selection}
\alias{add_codes_to_selection}
\title{Adds codes surrounding the selected text}
\usage{
add_codes_to_selection(selection, codes)
}
\arguments{
\item{selection}{The selection of text to be coded}

\item{codes}{The code or codes to be added to the document}
}
\description{
Adds codes surrounding the selected text
}
