% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qcodes.R
\name{add_discovered_code}
\alias{add_discovered_code}
\title{Update codes data frame
Add discovered codes to the codes data frame}
\usage{
add_discovered_code(codes_list = "", code_data_frame = NULL,
  save_path = "")
}
\arguments{
\item{codes}{A list of codes (usually from a coded document)}

\item{codes_data_frame}{Existing data frame of QCODE codes}
}
\description{
Update codes data frame
Add discovered codes to the codes data frame
}
