% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qcodes.R
\name{add_new_code}
\alias{add_new_code}
\title{Update codes data frame with description
Add code to the codes data frame}
\usage{
add_new_code(new_code = "", description = "", code_data_frame = NULL,
  codes_df_path = "")
}
\arguments{
\item{new_code}{The name of a code (usually from a add code form)}

\item{code_data_frame}{Existing data frame of QCODE codes}

\item{codes_df_path}{The path where the updated code data frame should be saved}
}
\description{
Update codes data frame with description
Add code to the codes data frame
}
