% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogmap.R
\name{add_new_coding}
\alias{add_new_coding}
\title{Update codings data frame with description
Add coding to the codings data frame}
\usage{
add_new_coding(doc_id, concept_from, concept_to, coding_sign,
  coding_weight, coding_class, document_part, selected_text,
  coding_data_frame = NULL, codings_df_path = "")
}
\arguments{
\item{coding_data_frame}{Existing data frame of QCODER codings}

\item{codings_df_path}{The path where the updated coding data frame should be saved}

\item{new_coding}{The name of a coding (usually from a add coding form)}
}
\description{
Update codings data frame with description
Add coding to the codings data frame
}
