% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogmap.R
\name{add_new_concept}
\alias{add_new_concept}
\title{Update concepts data frame with description
Add concept to the concepts data frame}
\usage{
add_new_concept(new_concept = "", description = "",
  concept_data_frame = NULL, concepts_df_path = "")
}
\arguments{
\item{new_concept}{The name of a concept (usually from a add concept form)}

\item{concept_data_frame}{Existing data frame of QCONCEPT concepts}

\item{concepts_df_path}{The path where the updated concept data frame should be saved}
}
\description{
Update concepts data frame with description
Add concept to the concepts data frame
}
