% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_files.R
\name{build_empty_paths}
\alias{build_empty_paths}
\title{Build the paths for empty files}
\usage{
build_empty_paths(project_name, data_frame_name = "",
  df_path = "data_frames", project_path = "")
}
\arguments{
\item{project_name}{Name of the project}

\item{data_frame_name}{Name of the data frame that will contain the data}

\item{df_path}{path segment(s) to the created data frame file
from the project path}

\item{project_path}{Path to the project (not including project_name).}
}
\description{
Build the paths for empty files
}
