% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_files.R
\name{build_paths}
\alias{build_paths}
\title{Build the paths for file creation}
\usage{
build_paths(
  project_name,
  data_path = "",
  data_frame_name = "",
  df_path = "data_frames",
  project_path = ""
)
}
\arguments{
\item{project_name}{Name of the project}

\item{data_path}{Path segment to the data. The format for this may
depend on the function using the paths.}

\item{data_frame_name}{Name of the data frame that will contain the data}

\item{df_path}{path segment(s) to the created data frame file
from the project path}

\item{project_path}{Path to the project (not including project_name). This
will be set to getwd() if a value of "" is passed in.}
}
\value{
A named list of paths. "data_frame_path" is the path to the data frame
and "data" is the path to the data.
}
\description{
Builds the paths to the data to be imported and to the data
frame where the imported data is to be stored.
The project name is required, all other parameters may be set.
These each represent a segment of the path to a file.
If a project path is not set or set to "" it will be set to
the current working directory via getwd().
The Shiny qcode application assumes that the data frame folder will
be "data_frames" and that any new documents to be imported will
be in a folder called "documents".
}
