% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_files.R
\name{create_empty_unit_doc_file}
\alias{create_empty_unit_doc_file}
\title{Define an empty many to many unit to document map}
\usage{
create_empty_unit_doc_file(file_path = "data_frames",
  data_frame_name = "unit_document_map", project_name)
}
\arguments{
\item{file_path}{Path to store data frame on.}

\item{data_frame_name}{Name of the data frame that will contain the map}

\item{project_name}{Name of the project, if it exists}
}
\description{
Define an empty many to many unit to document map
}
\examples{
create_qcoder_project(project_name = "_my_qcoder_project")
create_empty_unit_doc_file(project_name = "_my_qcoder_project")
unlink("./_my_qcoder_project", recursive=TRUE)
}
