% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_files.R
\name{create_empty_units_file}
\alias{create_empty_units_file}
\title{Define an empty many units data frame}
\usage{
create_empty_units_file(project_name, data_frame_name = "qcoder_unit",
  df_path = "data_frames", project_path = "")
}
\arguments{
\item{project_name}{Name of the project}

\item{data_frame_name}{Name of the data frame that will contain the data}

\item{df_path}{Path segment(s) to the created data frame}

\item{project_path}{Path to the project (not including project name.}
}
\description{
Define an empty many units data frame
}
\examples{
create_qcoder_project(project_name = "_my_qcoder_project")
create_empty_units_file(project_name = "_my_qcoder_project")
unlink("./_my_qcoder_project", recursive=TRUE)
}
