% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_qcoder_project}
\alias{create_qcoder_project}
\title{Create a standard set of folders for a QCoder project}
\usage{
create_qcoder_project(project_name, sample = FALSE)
}
\arguments{
\item{project_name}{A string project name to be located in the
current working directory or a path to a project folder.}

\item{sample}{Logical that indicates that the sample data should be copied to the project.}
}
\description{
Create a standard set of folders for a QCoder project
}
\examples{
create_qcoder_project(project_name = "my_qcoder_project")
}
