% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_files.R
\name{create_unit_doc_file}
\alias{create_unit_doc_file}
\title{Define a many to many unit to document map}
\usage{
create_unit_doc_file(file_path, data_frame_name)
}
\description{
Define a many to many unit to document map
}
