% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_update_document}
\alias{do_update_document}
\title{Update document
Updates the text field of the documents data frame, typically after
pressing Save button in the Shiny App.  May
also be used in the console.}
\usage{
do_update_document(updated, docs_df_path, this_doc_path)
}
\arguments{
\item{updated}{The updated text as a character string}

\item{docs_df_path}{Location of the documents rds file.}

\item{this_doc_path}{Name of record to be updated, as recorded in "doc_path"
field of data frame.}
}
\description{
Update document
Updates the text field of the documents data frame, typically after
pressing Save button in the Shiny App.  May
also be used in the console.
}
\examples{

}
