% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qcodes.R
\name{get_codes}
\alias{get_codes}
\title{Extract codes from text
Take coded text and extract the codes, assuming they are correctly formatted.}
\usage{
get_codes(doc_text)
}
\arguments{
\item{doc_text}{The text data for a single document}
}
\description{
Extract codes from text
Take coded text and extract the codes, assuming they are correctly formatted.
}
\examples{
unlink("./my_qcoder_project", recursive=TRUE)
}
