% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_files.R
\name{import_project_data}
\alias{import_project_data}
\title{Read data into a project
Convenience method to read raw data from standard locations and using
standard names in a project folder structure.}
\usage{
import_project_data(project_name)
}
\arguments{
\item{project_name}{The project name. This should represent the folder
holding the project.}
}
\description{
Read data into a project
Convenience method to read raw data from standard locations and using
standard names in a project folder structure.
}
