% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qcodes.R
\name{parse_qcodes}
\alias{parse_qcodes}
\title{Parse coded text}
\usage{
parse_qcodes(x, ...)
}
\arguments{
\item{x}{A data frame containing the text to be coded; requires columns "doc_id" and "document_text"}

\item{...}{Other parameters optionally passed in}
}
\description{
Take a text document containing coded text of the form:
"stuff to ignore (QCODE) coded text we care about (/QCODE){#qcode} more stuff to ignore"
and turn it into a dataframe with one row per coded item, of the form:
docid,qcode,text
}
\details{
Replaces newline characters with an HTML "br" in the captured text
returns an empty dataframe (no rows) if no qcodes were found.
}
\examples{
unlink("./_my_qcoder_project", recursive=TRUE)
}
