% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qcodes.R
\name{parse_splititem}
\alias{parse_splititem}
\title{Parse a single item within a document}
\usage{
parse_splititem(splititem, df, doc_id, dots)
}
\arguments{
\item{splititem}{String usually generated by parse_one_document()}

\item{df}{Data frame for storing parsed data}

\item{doc_id}{The doc_id for the document this string is part of.}

\item{dots}{List of additional options passed in.}
}
\description{
Parse a single item within a document
}
