% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qcode}
\alias{qcode}
\title{This launches the coder Shiny app}
\usage{
qcode(use_wd = TRUE)
}
\arguments{
\item{use_wd}{Whether or not the current working directory when launching
qcoder should be used as the base from which the project file is selected.}
}
\description{
This launches the coder Shiny app
}
\examples{
if (interactive()) {
 qcode()
}
}
