% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qcode_custom}
\alias{qcode_custom}
\title{This launches the coder custom Shiny app}
\usage{
qcode_custom()
}
\description{
This launches the coder custom Shiny app
}
\examples{
\dontrun{
  qcode_custom()
}
}
