% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcoder-package.R
\docType{package}
\name{qcoder-package}
\alias{qcoder-package}
\alias{qcoder}
\title{Code Qualitative Data
A light weight approach to qualitative coding and analysis}
\description{
Light weight coding
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropenscilabs/qcoder}
  \item Report bugs at \url{https://github.com/ropenscilabs/qcoder/issues}
}

}
\author{
\strong{Maintainer}: Elin Waring \email{elin.waring@gmail.com}

Authors:
\itemize{
  \item Beth Duckles \email{bduckles@gmail.com}
  \item Dan Sholler \email{sholler.daniel@gmail.com}
  \item Jenny Draper \email{learithe@gmail.com}
  \item Ted Laderas \email{roshioshi@gmail.com}
}

Other contributors:
\itemize{
  \item Fred VanWindekens \email{f.vanwindekens@gmail.com} [contributor]
  \item Yuri Dias \email{yurimathe.yp@gmail.com} [contributor]
  \item Sheriff Colley \email{rikunaruto0198@gmail.com} [contributor]
  \item Jamal Jones \email{Jones-Jamal@outlook.com} [contributor]
  \item Michael Lieberman \email{MichaelAYonkers@gmail.com} [contributor]
  \item Tenzing Sherpa \email{tsherpa2016@gmail.com} [contributor]
  \item Stephanie Guillou \email{stephane.guillou@member.fsf.org} [contributor]
}

}
\keyword{internal}
