% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogmap.R
\name{read_concept_data}
\alias{read_concept_data}
\title{Create a file of concepts from csv file
Use this is you have a spreadsheet of concepts already created.}
\usage{
read_concept_data(file_path = "concepts/concepts.csv",
  concepts_df_path = "", data_frame_name = "qcoder_concepts",
  project_name, project_path = "")
}
\arguments{
\item{file_path}{Path to a file containing  concept data in csv.}

\item{concepts_df_path}{Full path to the concepts data frame.}

\item{data_frame_name}{The name of the RDS file that the data frame will be stored in.}

\item{project_name}{Name of the project, which matches folder name}

\item{project_path}{Full path to the project folder}
}
\description{
Create a file of concepts from csv file
Use this is you have a spreadsheet of concepts already created.
}
\examples{
 \dontrun{
fp <-"inst/example_concepts/"
dfn <- "test_concepts"
read_concept_data(fp, dfn)
}
}
