% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_files.R
\name{read_unit_document_map_data}
\alias{read_unit_document_map_data}
\title{Create a data frame of unit to document links from csv file
Use this is you have a spreadsheet already created.}
\usage{
read_unit_document_map_data(file_path = "unit_document_map.csv",
  data_frame_name = "qcoder_unit_document_map", project_name,
  project_path = "", units_docs_path = "")
}
\arguments{
\item{file_path}{path to a file containing  unit document map data in csv.}

\item{data_frame_name}{The name of the RDS file that the data frame will be stored in.}

\item{project_name}{Name of project if available}

\item{project_path}{Full path to the project folder}

\item{units_docs_path}{Full path to the documents data frame.}
}
\description{
Create a data frame of unit to document links from csv file
Use this is you have a spreadsheet already created.
}
\examples{
create_qcoder_project(project_name = "my_qcoder_project", sample = TRUE)
project_name = "my_qcoder_project"
fp <-"unit_document_map.csv"
dfn <- "qcoder_unit_document_map"
read_unit_document_map_data(fp, dfn, project_name = "my_qcoder_project")
unlink("./my_qcoder_project", recursive=TRUE)
}
