% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{txt2html}
\alias{txt2html}
\title{Format text as HTML
Minimal conversion of a text to html}
\usage{
txt2html(text)
}
\arguments{
\item{text}{text to be converted}
}
\description{
Format text as HTML
Minimal conversion of a text to html
}
\examples{
txt2html("The quick brown (QCODE)fox(/QCODE){#animal} jumped over ")
}
