% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_project}
\alias{validate_project}
\title{Check for a valid qcoder project}
\usage{
validate_project(path_to_test)
}
\arguments{
\item{path_to_test}{Path to possible project folder}
}
\value{
NULL for valid project, Error otherwise.
}
\description{
Check for a valid qcoder project
}
\examples{
create_qcoder_project(project_name = "_my_qcoder_project")
validate_project("_my_qcoder_project")
unlink("./_my_qcoder_project", recursive=TRUE)
}
