% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_project_files}
\alias{validate_project_files}
\title{Check for required imported data frames.}
\usage{
validate_project_files(path_to_test)
}
\arguments{
\item{path_to_test}{Path to possible project folder}
}
\value{
NULL for valid project, Error otherwise.
}
\description{
Check for required imported data frames.
}
\examples{
create_qcoder_project(project_name = "_my_qcoder_project", sample = TRUE)
import_project_data("_my_qcoder_project")
validate_project_files("_my_qcoder_project")
unlink("./_my_qcoder_project", recursive=TRUE)
}
