% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qcodes.R
\name{add_discovered_code}
\alias{add_discovered_code}
\title{Update codes data frame
Add discovered codes to the codes data frame}
\usage{
add_discovered_code(codes_list = "", code_data_frame = NULL,
  codes_df_path = "")
}
\arguments{
\item{codes_list}{A list of codes (usually from a coded document)}

\item{code_data_frame}{Existing data frame of QCODE codes}

\item{codes_df_path}{The path where the updated code data frame should be saved}
}
\description{
Update codes data frame
Add discovered codes to the codes data frame
}
