% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_files.R
\name{add_new_documents}
\alias{add_new_documents}
\title{Add new documents
Adds new document or documents to an existing documents data frame.}
\usage{
add_new_documents(files, file_path = "", docs_df_path)
}
\arguments{
\item{files}{file tibble produced by ShinyFiles}

\item{file_path}{Full path to the data set of documents}

\item{docs_df_path}{Existing data frame of text documents}
}
\description{
Add new documents
Adds new document or documents to an existing documents data frame.
}
