% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_unit}
\alias{add_unit}
\title{Add unit
Append a new unit record to the existing data frame}
\usage{
add_unit(units_df, new_unit, units_df_path)
}
\arguments{
\item{units_df}{Existing units data frame}

\item{new_unit}{text name of a new unit (single name only)}

\item{units_df_path}{full path to the units data frame}
}
\description{
Add unit
Append a new unit record to the existing data frame
}
\examples{
unlink("./_my_qcoder_project", recursive=TRUE)
}
