% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_files.R
\name{create_empty_code_file}
\alias{create_empty_code_file}
\title{Create an empty codes data set}
\usage{
create_empty_code_file(path)
}
\arguments{
\item{path}{Full path to data frame to be created.}
}
\description{
Used to create a codes data frame with no data but that can
have data added. File is placed in the data_frames folder.
}
\examples{
create_qcoder_project(project_name = "_my_qcoder_project")
path <- file.path(getwd(),
  "_my_qcoder_project/data_frames/qcoder_codes__my_qcoder_project")
create_empty_docs_file(path)
unlink("./_my_qcoder_project", recursive=TRUE)
}
