% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_files.R
\name{create_empty_units_file}
\alias{create_empty_units_file}
\title{Define an empty  units data frame}
\usage{
create_empty_units_file(path)
}
\arguments{
\item{path}{Full path to data frame to be created.}
}
\description{
Define an empty  units data frame
}
\examples{
create_qcoder_project(project_name = "_my_qcoder_project")
path <- file.path(getwd(),
  "_my_qcoder_project/data_frames/qcoder_units__my_qcoder_project")
create_empty_docs_file(path)
unlink("./_my_qcoder_project", recursive=TRUE)
}
