% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qcodes.R
\name{error_check}
\alias{error_check}
\title{Check for coding errors
Checks the current document for coding errors.}
\usage{
error_check(document)
}
\arguments{
\item{document}{The document to be scanned for errors.}
}
\description{
Check for coding errors
Checks the current document for coding errors.
}
\examples{

}
