% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{make_js_listener}
\alias{make_js_listener}
\title{Makes a bit of javascript that lets us listen in on the selection}
\usage{
make_js_listener(editor_name, position)
}
\arguments{
\item{editor_name}{The name given the editor}

\item{position}{The starting row and column location}
}
\value{
javascript text
}
\description{
Makes a bit of javascript that lets us listen in on the selection
}
\examples{

}
