% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qcodes.R
\name{parse_qcodes}
\alias{parse_qcodes}
\title{Take a text document containing coded text of the form:
"stuff to ignore (QCODE) coded text we care about (/QCODE){#qcode} more stuff to ignore"
and turn it into a dataframe with one row per coded item, of the form:
docid,qcode,text}
\usage{
parse_qcodes(x, ...)
}
\arguments{
\item{x}{A data frame containing document texts}

\item{...}{Other parameters optionally passed in}
}
\description{
Replaces newline characters with "<br>" in the captured text
}
