% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcoder-package.R
\docType{"_PACKAGE}
\name{qcoder-package}
\alias{qcoder-package}
\alias{qcoder}
\title{Code Qualitative Data
A light weight approach to qualitative coding and analysis}
\description{
Light weight coding
}
