% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_files.R
\name{read_code_data}
\alias{read_code_data}
\title{Create a file of codes from csv file
Use this if you have a spreadsheet of codes already created.}
\usage{
read_code_data(
  project_name,
  data_path = "codes/codes.csv",
  df_path = "data_frames",
  data_frame_name = "qcoder_codes",
  project_path = ""
)
}
\arguments{
\item{project_name}{Name of the project, which matches folder name}

\item{data_path}{Path to a file containing  code data in csv.}

\item{df_path}{Full path to the codes data frame.}

\item{data_frame_name}{The name of the RDS file that the data frame
will be stored in.}

\item{project_path}{Full path to the project folder}
}
\description{
Create a file of codes from csv file
Use this if you have a spreadsheet of codes already created.
}
\examples{
create_qcoder_project(project_name = "_my_qcoder_project", sample = TRUE)
read_code_data(project_name = "_my_qcoder_project")
unlink("./_my_qcoder_project", recursive=TRUE)
}
