% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_files.R
\name{read_raw_data}
\alias{read_raw_data}
\title{Create a data frame of documents}
\usage{
read_raw_data(folder_path = "/documents/",
  data_frame_name = "qcoder_documents", project_name = NULL,
  docs_df_path = "", project_path = "")
}
\arguments{
\item{folder_path}{path to a folder contain text files to be analyzed.}

\item{data_frame_name}{The name of the RDS file that the data frame will be stored in.}

\item{project_name}{Name of the Qcoder project}

\item{docs_df_path}{Full path to the docs data frame.}

\item{project_path}{Full path to the project folder.}
}
\description{
Create a data frame of documents
}
\examples{
 \dontrun{
fp <-"/documents/"
dfn <- "testdata"
read_raw_data(fp, dfn, "")
}
}
