% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_files.R
\name{read_unit_document_map_data}
\alias{read_unit_document_map_data}
\title{Create a data frame of unit to document links from csv file
Use this is you have a spreadsheet already created.}
\usage{
read_unit_document_map_data(file_path = "unit_document_map.csv",
  data_frame_name = "qcoder_unit_document_map", project_name)
}
\arguments{
\item{file_path}{path to a file containing  unit document map data in csv.}

\item{data_frame_name}{The name of the RDS file that the data frame will be stored in.}

\item{project_name}{Name of project if available}
}
\description{
Create a data frame of unit to document links from csv file
Use this is you have a spreadsheet already created.
}
\examples{
 \dontrun{
fp <-"units/units.csv"
dfn <- "my_units"
read_unit_data(fp, dfn)
}
}
