% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_links}
\alias{update_links}
\title{Update document to unit links
Saves or updates the links between observation units and documents}
\usage{
update_links(checked = "", docs_df_path = "", this_doc_path = "",
  units_docs_path = "")
}
\arguments{
\item{checked}{vector of new or updated links}

\item{docs_df_path}{full path to document dataset}

\item{this_doc_path}{value of doc_path for the document}

\item{units_docs_path}{full path of the data frame of unit to docs links}
}
\description{
Update document to unit links
Saves or updates the links between observation units and documents
}
\examples{

unlink("./_my_qcoder_project", recursive=TRUE)
}
